==> scripts/global/00a_eez_polygon_wrangling.Rmd <==
---
output:
  html_document: default

title: "Distant Water Fishing Atlas - EEZ Shapefile Wrangling"
author: "Kat Millage, Matt Warham"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
---

# Introduction

This script wrangles the EEZ GeoDatabase file used in the tool. A key part of the later connectivity analysis is having central coordinates applied to each polygon, so that is done here as well. This relies upon the following data sources: 
- Flanders Marine Institute (2019). Maritime Boundaries Geodatabase: Maritime Boundaries and Exclusive Economic Zones (200NM), version 11. Available online at https://www.marineregions.org/. https://doi.org/10.14284/386


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

### Load packages ------
library(sf) # spatial manipulation
library(here) # path names
library(knitr) # knit document
library(countrycode) # country names 
library(janitor)
library(mapview)
library(rmapshaper)
library(tidyverse) # data manipulation

### Create results directories
data_dir <- here::here("data/edited/")
  if (dir.exists(data_dir) == F) { dir.create(data_dir, recursive = T) }

results_dir <- here::here("results/global/00a-eez-polygon-wrangling/")
  if (dir.exists(results_dir) == F) { dir.create(results_dir, recursive = T) }

### Also save to app data directory
app_dir <- here::here("app-global/data/")
  if (dir.exists(app_dir) == F) { dir.create(app_dir, recursive = T) }
```

# Marine Regions EEZs (v11) GeoPackage

This datafile is freely available from https://www.marineregions.org/downloads.php. For this analysis, we're using the World EEZ v11 (2019-11-18, 127 MB) GeoPackage. Due to size, this file is not contained in this repository (but a simplified version is). 

## Simplify polygons

We first simplify the EEZ polygons because they are much higher resolution than we need for visualization purposes. This code block is commented out because the raw data is not contained in this repository due to size constraints, and this process is time consuming. The simplified version is available in this repository and is instead read in below. 

```{r}
# Path where the raw data file lives - NOT IN THIS REPOSITORY
# eez_path <- "/Volumes/GoogleDrive/Shared\ drives/emlab/data/marine-regions-eez-v11/World_EEZ_v11_20191118_gpkg/eez_v11.gpkg"
# 
# eez_sf <- st_read(eez_path) %>%
#   clean_names()
# 
# # Simplify polygons to make working with them easier and faster - we don't need high resolution here as we're just using these for visualization purposes
# eez_sf_simple <- st_make_valid(rmapshaper::ms_simplify(eez_sf, keep_shapes = TRUE, keep = 0.01))
# 
# # Save simplified object so we have a version of the data in this repository
# st_write(eez_sf_simple, paste0(data_dir, "world_eez_v11_simple.gpkg"), delete_layer = T)

```

```{r}
eez_sf_simple <- st_read(paste0(data_dir, "world_eez_v11_simple.gpkg")) %>%
  st_set_crs(4326)
```

## Create expanded version of polygons (-360 to 360 degrees)

Because we are going to center different pages of our app on different regions, we want to make an duplicated polygon object that ranges from -360 to 360 degrees. 

```{r}
# Make left duplicate (-540, -180)
eez_sf_left <- (st_geometry(eez_sf_simple) - c(360, 0)) %>%
  st_set_crs(4326)

# Make right duplicate (180 - 540)
eez_sf_right <- (st_geometry(eez_sf_simple) + c(360, 0)) %>%
  st_set_crs(4326)

# Assign geometry
eez_sf_left <- st_set_geometry(eez_sf_simple, eez_sf_left)
eez_sf_right <- st_set_geometry(eez_sf_simple, eez_sf_right)

# Combine
eez_sf_all <- eez_sf_simple %>%
  rbind(eez_sf_left) %>% 
  rbind(eez_sf_right) %>%
  st_make_valid() %>%
  group_by(mrgid) %>%
  summarize(a = 1) %>%
  ungroup() %>%
  dplyr::select(-a)

# Crop
eez_sf_crop <- eez_sf_all %>%
  st_crop(xmin = -360, ymin = -90, xmax = 360, ymax = 90)

# Keep only polygons
eez_sf_polygon <- st_collection_extract(
  eez_sf_crop,
  type = c("POLYGON"),
  warn = FALSE
)

ggplot() + geom_sf(data = eez_sf_polygon, aes(fill = mrgid))

# Save
st_write(eez_sf_polygon, paste0(data_dir, "world_eez_v11_simple_neg360_360.gpkg"), delete_layer = T)
```

## EEZ Attributes

```{r}
# Extract attributes we want to keep
eez_attrib <- eez_sf_simple %>%
  st_drop_geometry()

write_csv(eez_attrib, paste0(results_dir, "eez_attributes_raw.csv"))

# Now let's do some editing so we have one row for each territory (dealing with joing regime and disputed areas)
eez_attrib_all <- eez_attrib %>%
  rowwise() %>%
  mutate(iso_ter_all = list(c(iso_ter1, iso_ter2, iso_ter3)),
         iso_sov_all = list(c(iso_sov1, iso_sov2, iso_sov3))) %>%
  dplyr::select(mrgid, pol_type, geoname, iso_ter_all, iso_sov_all)

# Assign country names
eez_attrib_expanded <- eez_attrib_all %>%
  unnest(cols = c("iso_ter_all", "iso_sov_all")) %>%
  mutate(iso_ter = ifelse(!is.na(iso_ter_all), iso_ter_all, iso_sov_all)) %>%
  dplyr::filter(!is.na(iso_ter)) %>%
  mutate(name_ter = countrycode(iso_ter, "iso3c", 'country.name.en'),
         name_sov = countrycode(iso_sov_all, "iso3c", 'country.name.en'),
         region = countrycode(iso_ter, "iso3c", "region"))

# Manually assign regions for those that are missing or are incorrect
eez_attrib_expanded <- eez_attrib_expanded %>%
  mutate(region_custom = case_when(iso_ter == "ATA" ~ "NA",
                                   iso_ter == "ATF" ~ "Sub-Saharan Africa",
                                   iso_ter == "ESH" ~ "Middle East & North Africa",
                                   iso_ter == "MYT" ~ "Sub-Saharan Africa",
                                   iso_ter == "REU" ~ "Sub-Saharan Africa",
                                   iso_ter == "SHN" ~ "Sub-Saharan Africa",
                                   iso_ter == "SJM" ~ "Europe & Central Asia",
                                   iso_ter == "UMI" ~ "East Asia & Pacific",
                                   iso_ter == "WLF" ~ "East Asia & Pacific",
                                   iso_ter == "SGS" ~ "Latin America & Caribbean", # Fix South Georgia and the South Sandwich Islands
                                   iso_ter == "MLT" ~ "Europe & Central Asia", # Fix Malta
                                   TRUE ~ region)) %>%
  dplyr::select(-region, -iso_ter_all) %>%
  rename(iso_sov = iso_sov_all,
         region = region_custom)

# Make some manual corrections for certain disputed/overlapping/joint areas so they're attributed to each claiming state
eez_attrib_expanded_fix <- eez_attrib_expanded %>%
  dplyr::filter(pol_type != "200NM") %>%
  arrange(geoname) %>%
  mutate(pol_type_new = case_when(pol_type == "200 NM" ~ "200NM",
                                  mrgid == 5695 ~ "200NM",
                                  TRUE ~ pol_type),
         geoname_new = case_when(mrgid == 5695 ~ "Ukrainian Exclusive Economic Zone",
                                 TRUE ~ geoname),
         name_ter_new = case_when(iso_ter == "FRO" ~ "Denmark",
                                  iso_ter == "SJM" ~ "Norway",
                                  iso_ter == "FLK" & iso_sov == "GBR" ~ "United Kingdom",
                                  iso_ter == "FLK" & iso_sov == "ARG" ~ "Argentina",
                                  iso_ter == "FRA" & mrgid == 48946 ~ "Réunion",
                                  iso_ter == "FRA" & mrgid == 48948 ~ "New Caledonia",
                                  iso_ter == "MYT" & iso_sov == "COM" ~ "Comoros",
                                  iso_ter == "MYT" & iso_sov == "FRA" ~ "French Southern Territories",
                                  iso_ter == "SGS" ~ name_sov,
                                  iso_ter == "ESH" ~ name_sov,
                                  TRUE ~ name_ter),
         iso_ter_new = case_when(iso_ter == "FRO" ~ "DNK",
                                 iso_ter == "SJM" ~ "NOR",
                                 iso_ter == "FLK" & iso_sov == "GBR" ~ "GBR",
                                 iso_ter == "FLK" & iso_sov == "ARG" ~ "ARG",
                                 iso_ter == "FRA" & mrgid == 48946 ~ "REU",
                                 iso_ter == "FRA" & mrgid == 48948 ~ "NCL",
                                 iso_ter == "MYT" & iso_sov == "COM" ~ "COM",
                                 iso_ter == "MYT" & iso_sov == "FRA" ~ "ATF",
                                 iso_ter == "SGS" ~ iso_sov,
                                 iso_ter == "ESH" ~ iso_sov,
                                 TRUE ~ iso_ter)) %>%
  dplyr::select(-pol_type, -geoname, -name_ter, -iso_ter) %>%
  rename(pol_type = pol_type_new, geoname = geoname_new, name_ter = name_ter_new, iso_ter = iso_ter_new) %>%
  dplyr::filter(iso_ter != "ESH" & iso_sov != "ESH") # remove independent entry for Western Sahara

# Modify some regions again so that when one of the administering territories is in a different region, we can see it.... 
eez_attrib_expanded_fix_again <- eez_attrib_expanded_fix %>%
  mutate(region = case_when(mrgid == 8383 & iso_ter == "GBR" ~ "Europe & Central Asia", # SGS
                            mrgid == 8389 & iso_ter == "GBR" ~ "Europe & Central Asia", # Falkland Islands
                            mrgid == 48946 & iso_ter == "REU" ~ "Sub-Saharan Africa", # Ile Tromelin
                            mrgid == 48948 & iso_ter == "NCL" ~ "East Asia & Pacific", # Matthew + Hunter
                            TRUE ~ region))

# Join back together
eez_attrib_expanded <- eez_attrib_expanded %>%
  dplyr::filter(pol_type == "200NM") %>%
  bind_rows(eez_attrib_expanded_fix_again)
                                 
write_csv(eez_attrib_expanded, paste0(results_dir, "eez_attributes_expanded.csv"))

# Now let's condense the attributes so there's a single row for each zone 
eez_attrib_flat <- eez_attrib_expanded %>%
  group_by(mrgid) %>%
  summarize(pol_type = unique(pol_type),
            geoname = unique(geoname),
            iso_sov = list((unique(iso_sov))),
            name_sov = list((unique(name_sov))),
            iso_ter = list((unique(iso_ter))),
            name_ter = list((unique(name_ter))),
            region = list((unique(region))))

save(eez_attrib_flat, file = paste0(results_dir, "eez_attributes_flat.rda"))
```

## Duplicate Polygons

Since many of our polygons have multiple administering territories (e.g., joint regime and disputed areas), we want to duplicate these so we have a polygon associated with that area assigned to each territory. 

```{r}
# Add attributes back to our polygons
eez_sf_out <- eez_sf_polygon %>%
  full_join(eez_attrib_expanded, by = c("mrgid")) %>%
  dplyr::filter(!is.na(geoname))
```

## Find Centroids 

In order to make our connectivity plots later on, we will need the centroids of each EEZ (and flag state) so we can create the connectivity lines. In the Distant Water Fishing Atlas, this gets a little tricky because we want our plots to be centered on each EEZ of interest, depending on the region. 

This is why we duplicated and extended our polygons such that the total extent ranges from -360 to 360 degrees. We can then clip these to the appropriate extents for each of our seven regions, such that no polygons (or connectivity lines) should be getting bifurcated. 

For our seven regions, we're using the following regional centroids: 
- East Asia & Pacific: (170, -5)
- Europe & Central Asia: (0, 50)
- Latin America & Caribbean: (-70, -15)
- Middle East & North Africa: (25, 10)
- North America: (-100, 40)
- South Asia: (80, 15)
- Sub-Saharan Africa: (25, -15) 

As a starting point, we'll therefore have our map extents set by going +/- 180 degrees from those points: 

- East Asia & Pacific: (-10, 350)
- Europe & Central Asia: (-180, 180)
- Latin America & Caribbean: (-250, -110)
- Middle East & North Africa: (-155, 205)
- North America: (-280, 80)
- South Asia: (-100, 260)
- Sub-Saharan Africa: (-155, 205) 

For simplicity, we should be able to get everthing we need by using 0 - 360 for the East Asia & Pacific Region, and -180 - 180 for everything else. 

```{r}
map_coord <- list(east_asia_pacific = c(0, 360),
                  all_others = c(-180, 180))
```

```{r}
  # East Asia & Pacific
  east_asia_pacific_eez_sf <- st_crop(eez_sf_out, c(xmin = map_coord$east_asia_pacific[1],
                                                xmax = map_coord$east_asia_pacific[2],
                                                ymin=-90, ymax=90))
  east_asia_pacific_eez_centroids <- east_asia_pacific_eez_sf %>%
    st_centroid() %>%
    dplyr::select(mrgid, pol_type, geoname, eez_ter_iso3 = iso_ter, eez_ter_name = name_ter, eez_sov_iso3 = iso_sov, eez_sov_name = name_sov, region) %>%
    dplyr::filter(region == "East Asia & Pacific")
  
  # All other regions
  all_others_eez_sf <- st_crop(eez_sf_out, c(xmin = map_coord$all_others[1],
                                                xmax = map_coord$all_others[2],
                                                ymin=-90, ymax=90))
  all_other_eez_centroids <- all_others_eez_sf %>%
    st_centroid() %>%
    dplyr::select(mrgid, pol_type, geoname, eez_ter_iso3 = iso_ter, eez_ter_name = name_ter, eez_sov_iso3 = iso_sov, eez_sov_name = name_sov, region) %>%
    dplyr::filter(region != "East Asia & Pacific")

  # Bind together 
  eez_centroids <- east_asia_pacific_eez_centroids %>%
    rbind(all_other_eez_centroids)
  
  # Put centroids in lon/lat columns
  eez_centroid_coordinates <- tibble(mrgid = eez_centroids$mrgid,
                                     pol_type = eez_centroids$pol_type,
                                     geoname = eez_centroids$geoname,
                                     eez_ter_iso3 = eez_centroids$eez_ter_iso3,
                                     eez_ter_name = eez_centroids$eez_ter_name,
                                     eez_sov_iso3 = eez_centroids$eez_sov_iso3,
                                     eez_sov_name = eez_centroids$eez_sov_name,
                                     region = eez_centroids$region) %>% 
    bind_cols(do.call(rbind, st_geometry(eez_centroids)) %>%
                as_tibble() %>%
                setNames(c("eez_lon", "eez_lat"))) %>%
    arrange(eez_ter_iso3)
```

## Manual Corrections of Centroids

There might be a few centroids we want to correct - Do that here. 

```{r}
# Manual centroid corrections
eez_centroid_coordinates_corrected <- eez_centroid_coordinates 

```

## Add Centroids to polygons

```{r}
# Add Centroids
eez_sf_out <- eez_sf_out %>%
  rename(eez_ter_iso3 = iso_ter, eez_ter_name = name_ter, eez_sov_iso3 = iso_sov, eez_sov_name = name_sov) %>%
  left_join(eez_centroid_coordinates_corrected, by = c("mrgid", "pol_type", "geoname", "eez_ter_iso3", "eez_ter_name", "eez_sov_iso3", "eez_sov_name",  "region"))
```

## Save

```{r}
# Save
write_csv(eez_centroid_coordinates_corrected, paste0(results_dir, "eez_centroid_coordinates.csv"))
st_write(eez_sf_out, paste0(results_dir, "world_eez_v11_simple_neg360_360_centroids.gpkg"), delete_layer = T)
```

## Create region aggregated polygons for landing page

```{r}
# Regional aggregation
region_sf <- eez_sf_out %>%
  group_by(region) %>%
  summarize(geom = st_union(geom)) %>%
  dplyr::filter(region != "NA")
  
ggplot()+geom_sf(data = region_sf, aes(fill = region))

st_write(region_sf, paste0(data_dir, "world_eez_regions_neg360_360.gpkg"), delete_layer = T)
st_write(region_sf, paste0(results_dir, "world_eez_regions_neg360_360.gpkg"), delete_layer = T)
st_write(region_sf, paste0(app_dir, "world_eez_regions_neg360_360.gpkg"), delete_layer = T)
```

## Create territory aggregated polygons

Before we aggregate by territory, we also need to create an aggregated EU polygon. 

```{r}
### Aggregate EU Polygons
eu_countries <- c("AUT", "BEL", "BGR", "HRV", "CYP", "CZE", "DNK", "EST", "FIN", "FRA", "DEU", "GRC", "HUN", "IRL", "ITA", "LVA", "LTU", "LUX", "MLT", "NLD", "POL", "PRT", "ROU", "SVK", "SVN", "ESP", "SWE", "GBR")
  
eu_aggregated <- eez_sf_out %>%
  dplyr::filter((eez_ter_iso3 %in% eu_countries) & pol_type == "200NM") %>%
  mutate(eez_sov_iso3 = "EU") %>%
  group_by(eez_sov_iso3, pol_type) %>%
  summarize(eez_lon = mean(eez_lon, na.rm = T),
            eez_lat = mean(eez_lat, na.rm = T),
            mrgid = paste0(mrgid, collapse = ", "),
            geom = st_union(geom)) %>%
  ungroup() %>%
  mutate(region = "Europe & Central Asia",
         geoname_new = "Exclusive Economic Zone:\nEuropean Union",
         eez_ter_iso3 = "EU",
         eez_ter_name = "European Union",
         eez_sov_name = "European Union")
```


```{r}
# Territory aggregation (200nm areas only)
territory_sf <- eez_sf_out %>%
  dplyr::filter(pol_type == "200NM") %>%
  mutate(geoname_new = paste0("Exclusive Economic Zone:\n", eez_ter_name)) %>%
  group_by(region, eez_ter_name, eez_ter_iso3, eez_sov_name, eez_sov_iso3, pol_type, geoname_new) %>%
  summarize(eez_lon = mean(eez_lon, na.rm = T),
            eez_lat = mean(eez_lat, na.rm = T),
            mrgid = paste0(mrgid, collapse = ", "),
            geom = st_union(geom)) %>%
  ungroup()

others_sf <- eez_sf_out %>%
  dplyr::filter(pol_type != "200NM") %>%
  mutate(geoname_new = geoname) %>%
  dplyr::select(region, eez_ter_name, eez_ter_iso3, eez_sov_name, eez_sov_iso3, pol_type, geoname_new, eez_lon, eez_lat, mrgid, geom)

territory_sf_out <- territory_sf %>%
  rbind(others_sf) %>%
  rbind(eu_aggregated) %>%
  rename(eez_id = mrgid)
  
st_write(territory_sf_out, paste0(results_dir, "world_eez_ter_neg360_360_subsidy_atlas.gpkg"), delete_layer = T)
st_write(territory_sf_out, paste0(app_dir, "world_eez_ter_neg360_360_subsidy_atlas.gpkg"), delete_layer = T)
```
