==> scripts/global/01_data_wrangling.Rmd <==
---
output:
  html_document: default

title: "Distant Water Fishing Atlas - Data Wrangling (Subsidies)"
author: "Kat Millage, Matt Warham"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
---

```{r}
# Packages
library(knitr) # Necessary to convert file to .html
library(countrycode) # Country name matching
library(here) # File path handling
library(janitor) # Column name handling
library(tidyverse) # General data wrangling
library(readxl) # Load .xlsx files

# Code chunk defaults
knitr::opts_chunk$set(echo = FALSE, error = FALSE, message = FALSE, warning = FALSE, include = FALSE)

# Directories for the output files generated by this script
data_dir <- here::here("data", "edited/")
if (dir.exists(data_dir) == F) {
  dir.create(data_dir, recursive = T)
}

results_dir <- here::here("results", "global", "01-data-wrangling/")
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}
```

# Description

This script handles the preliminary data wrangling for the raw data sets underlying the Distant Water Fishing Atlas analysis. It outputs tidied data files to `r data_dir`. 

```{r}
### Paths to fisheries subsidies data files
sumaila_subsidies_path <- here::here("data", "raw", "Sumaila-et-al-2019-subsidies", "Subsidies_PEW_Final.xlsx")
schuhbauer_subsidies_ssf_path <- here::here("data", "raw", "Schuhbauer-et-al-2017-ssf-subsidies", "Subsidies_SSF_split_2018.csv")

### Lookup tables (manually created)
country_lookup_path <- here::here("data", "lookup-tables", "country_dependencies.csv")
sumaila_lookup_path <- here::here("data", "lookup-tables", "sumaila_subsidies_classification_lookup_table.csv")
schuhbauer_lookup_path <- here::here("data", "lookup-tables", "schuhbauer_subsidies_classification_lookup_table.csv")
```

The following data sources are included in this analysis: 

**Fisheries Subsidies Data**
Estimates of annual fisheries subsidies provision by state from Sumaila et al. (2019) in 2018 US\$. These data were provided to us directly by the authors and are stored at `r sumaila_subsidies_path`.
> Sumaila, U. R., Ebrahim, N., Schuhbauer, A., Skerritt, D., Li, Y., Kim, H. S., … Pauly, D. (2019). Updated estimates and analysis of global fisheries subsidies. Marine Policy, 109, 103695. https://doi.org/10.1016/j.marpol.2019.103695

Estimates of annual fisheries subsidies provision to small scale fisheries by state in 2018 US\$. Data used in this analysis are unpublished updates (based on the Sumaila et al. (2019) data) from those reported in Schuhbauer et al. (2017) (which are in 2009 US\$). These data were provided to us directly by the authors and are stored at `r schuhbauer_subsidies_ssf_path`. 
> Schuhbauer, A., Chuenpagdee, R., Cheung, W. W. L., Greer, K., & Sumaila, U. R. (2017). How subsidies affect the economic viability of small-scale fisheries. Marine Policy, 82, 114–121. https://doi.org/10.1016/j.marpol.2017.05.013

We also use the following tables to aid in our analysis: 

- A manually compiled table of overseas dependencies of WTO Member and Observer countries

- All subsidy types used by Sumaila et al. (2019)

# Note: Country naming, political entities, and dependencies

The naming of flag and coastal states varies widely across our data sources and we recognize that this can be a politically sensitive matter. For the purposes of merging multiple data sources, we therefore translate all state names from the raw data sources to ISO-3 character codes. Country names are only used for display purposes, and the use of any particular name is not meant to convey any options regarding sovereignty on behalf of the authors. 

```{r}
# Load country dependencies helper file
country_dependencies <- read_csv(country_lookup_path)

# Get all overseas dependencies that are relevant for the WTO - there are 49
territories <- country_dependencies %>%
  dplyr::filter(is_overseas_territory & is_WTO) %>%
  distinct(iso3, sovereign_iso3)

# Get all cooresponding sovereign states (we need to make sure any totals for these states include their dependencies) - there are 9
# Australia, Denmark, France, United Kingdom, Morocco, Netherlands, Norway, New Zealand, and the United States
sovereigns_with_territories <- territories %>%
  distinct(sovereign_iso3)

# EU countries
eu_countries <- country_dependencies$iso3[country_dependencies$is_EU]

# EU overseas territories
eu_territories <- country_dependencies$iso3[country_dependencies$is_overseas_territory & country_dependencies$sovereign_iso3 %in% eu_countries]
```

Additionally, it is important to note that the European Union as a political entity is a Member of the World Trade Organization, rather then the individual states that make up the European Union. We therefore need to create a profile for the European Union as a whole.  The following `r length(eu_country_names)` states are considered to be part of the European Union for the purposes of this analysis: `r paste(country_dependencies$WTO_name[country_dependencies$iso3 %in% eu_countries], collapse = ", ")`. 

The Netherlands, France, Denmark, and the United Kingdom also have overseas dependencies that should be considered to be a part of the European Union for statistical purposes.

Outside of the EU, the other dependencies we need to worry about are those that relate to the Austraila, Morocco, Norway, New Zealand, and the United States.

# Section 1 - Subsidies Data

## Estimates of global fisheries subsidies from Sumaila et al. (2019)

```{r}
### Data Wrangling
# Load and wrangle 2019 data
sumaila_2019_dat <- read_xlsx(sumaila_subsidies_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = case_when(country == "Dominican Rp" ~ "DOM",
                          country == "Micronesia" ~ "FSM",
                          country == "Untd Arab Em" ~ "ARE",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  dplyr::select(iso3, type = class, data_type, value = constant_2018_usd) %>%
  mutate(year = 2018,
         units = "2018 US$",
         source = "Sumaila et al. (2019)") %>%
  arrange(iso3)

# Load lookup table for classification scheme
sumaila_lookup <- read_csv(sumaila_lookup_path)

# Clean up the naming/spelling of the classification scheme as well as data type
sumaila_2019_dat_edit <- sumaila_2019_dat %>%
  left_join(sumaila_lookup, by = "type") %>%
  dplyr::select(iso3, year, category, category_name, type, type_name, value, data_type, units, source)

sumaila_2019_dat_edit$data_type[sumaila_2019_dat_edit$data_type == "not found evidence of subsidy"] <- "Found no evidence of subsidy"

# Deal with the EU
sumaila_2019_dat_eu <- sumaila_2019_dat_edit %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU",
         data_type = "Aggregate") %>%
  group_by(entity, year, category, category_name, type, type_name, data_type, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  rename(iso3 = entity)

# Add the EU back in
sumaila_2019_dat_edit <- sumaila_2019_dat_edit %>%
  bind_rows(sumaila_2019_dat_eu) %>%
  mutate(variable = "subsidies_Sumaila")

# Test to make sure there isn't data for any other dependencies - nope we're good
sumaila_2019_dat_overseas_territory <- sumaila_2019_dat_edit %>%
  dplyr::filter(iso3 %in% territories$iso3 & !(iso3 %in% eu_territories))

# Save
write_csv(sumaila_2019_dat_edit, paste0(data_dir, "sumaila_et_al_2019_subsidies_tidy.csv"))
write_csv(sumaila_2019_dat_edit, paste0(results_dir, "sumaila_et_al_2019_subsidies_tidy.csv"))

```

## Updated estimates of small-scale fisheries subsides from Schuhbauer et al. (2017)

Note: these data are unpublished updates (in 2018 US\$) of the data described in Schuhbauer et al. (2017). 

```{r}
# Load and wrangle data (2018 US$)
schuhbauer_subsidies_ssf_dat <- read_csv(schuhbauer_subsidies_ssf_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = case_when(country == "Dominican Rp" ~ "DOM",
                          country == "Micronesia" ~ "FSM",
                          country == "Untd Arab Em" ~ "ARE",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  mutate(prop_ssf = ssf_usd/constant_2018_usd) %>%
  dplyr::select(iso3, old_type_name = type, prop_ssf) %>%
  mutate(source = "Schuhbauer et al. (2017)") %>%
  arrange(iso3)
schuhbauer_subsidies_ssf_dat$prop_ssf[is.nan(schuhbauer_subsidies_ssf_dat$prop_ssf)] <- 0 # handling dividing by 0

# Load lookup table
schuhbauer_lookup <- read_csv(schuhbauer_lookup_path)

# Clean up the naming/spelling of the classification scheme as well as data type
schuhbauer_subsidies_ssf_dat_edit <- schuhbauer_subsidies_ssf_dat %>%
  left_join(schuhbauer_lookup, by = c("old_type_name")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, prop_ssf, source)

# Deal with the EU
schuhbauer_subsidies_ssf_dat_eu <- read_csv(schuhbauer_subsidies_ssf_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = case_when(country == "Dominican Rp" ~ "DOM",
                          country == "Micronesia" ~ "FSM",
                          country == "Untd Arab Em" ~ "ARE",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU") %>%
  group_by(entity, type) %>%
  summarize(constant_2018_usd = sum(constant_2018_usd, na.rm = T),
            ssf_usd = sum(ssf_usd, na.rm = T)) %>%
  ungroup() %>%
  mutate(prop_ssf = ssf_usd/constant_2018_usd) %>%
  dplyr::select(iso3 = entity, old_type_name = type, prop_ssf) %>%
  mutate(source = "Schuhbauer et al. (2017)") %>%
  left_join(schuhbauer_lookup, by = c("old_type_name")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, prop_ssf, source)

# Add the EU back in
schuhbauer_subsidies_ssf_dat_edit <- schuhbauer_subsidies_ssf_dat_edit %>%
  bind_rows(schuhbauer_subsidies_ssf_dat_eu)

# Test to make sure there isn't data for any other dependencies - nope we're good
schuhbauer_subsidies_ssf_dat_overseas_territory <- schuhbauer_subsidies_ssf_dat_edit %>%
  dplyr::filter(iso3 %in% territories$iso3 & !(iso3 %in% eu_territories))

# Save
write_csv(schuhbauer_subsidies_ssf_dat_edit, paste0(data_dir, "schuhbauer_ssf_subsidies_proportions.csv"))
write_csv(schuhbauer_subsidies_ssf_dat_edit, paste0(results_dir, "schuhbauer_ssf_subsidies_proportions.csv"))

```
