==> scripts/global/04_connectivity_analysis.Rmd <==
---
output:
  html_document: default

title: "Distant Water Fishing Atlas - Global Connectivity Analysis"
author: "Kat Millage, Matt Warham"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
---

# Introduction

This script uses our global list of distant water fishing vessels in 2018 to create the following summarized items: 
- Summary table showing total distant water fishing activity by EEZ/flag state 
- Connectivity lines for a given EEZ showing the origins of all distant water vessels

```{r setup, include=FALSE}
# Chunk options
knitr::opts_chunk$set(echo = TRUE)

### Load packages ------
library(sf) # spatial manipulation
library(lwgeom) # needed for curved lines
library(here) # path names
library(knitr) # knit document
library(countrycode) # country names 
library(tidyverse) # data manipulation
library(ggalt) # transform to robinson projection

# Path to distant water vessel list
vessel_list_dw_path <- here::here("results", "global", "03-distant-water-fishing", "vessel_list_2018_distant_water.csv")

# Path to high seas vessel list 
vessel_list_hs_path <- here::here("results", "global", "03-distant-water-fishing", "vessel_list_2018_high_seas.csv")

# Path to eez file
eez_centroid_path <- here::here("results", "global", "00a-eez-polygon-wrangling", "eez_centroid_coordinates.csv")

# Path to land file
land_centroid_path <- here::here("results", "global", "00b-land-polygon-wrangling", "land_centroid_coordinates.csv")

# Path to fao region file
fao_centroid_path <- here::here("results", "global", "00c-fao-region-polygon-wrangling", "fao_centroid_coordinates.csv")

### Create results directories
results_dir <- here::here("results/global/04-connectivity-analysis/")
  if (dir.exists(results_dir) == F) { dir.create(results_dir, recursive = T) }

### Also save to app data directory
app_dir <- here::here("app-global/data/")
  if (dir.exists(app_dir) == F) { dir.create(app_dir, recursive = T) }

### Also save to app data directory
app_www_dir <- here::here("app-global/www/")
  if (dir.exists(app_www_dir) == F) { dir.create(app_www_dir, recursive = T) }
```

# Setup

```{r}
# Load vessel list
vessel_list_dw_raw <- read.csv(vessel_list_dw_path, stringsAsFactors = F) 

# List of 2018 EU countries
eu_countries <- c("AUT", "BEL", "BGR", "HRV", "CYP", "CZE", "DNK", "EST", "FIN", "FRA", "DEU", "GRC", "HUN", "IRL", "ITA", "LVA", "LTU", "LUX", "MLT", "NLD", "POL", "PRT", "ROU", "SVK", "SVN", "ESP", "SWE", "GBR")
```

We first need to duplicate entries for areas with multiple administering territories so the stats are counted for each. 

```{r}
vessel_list_dw_expanded <- vessel_list_dw_raw %>%
  rowwise() %>%
  mutate(eez_ter_iso3_all = list(c(eez_ter1_iso3, eez_ter2_iso3, eez_ter3_iso3)),
         eez_sov_iso3_all = list(c(eez_sov1_iso3, eez_sov2_iso3, eez_sov3_iso3))) %>%
  dplyr::select(-eez_ter1_iso3, -eez_ter2_iso3, -eez_ter3_iso3, -eez_sov1_iso3, -eez_sov2_iso3, -eez_sov3_iso3) %>%
  unnest(cols = c("eez_ter_iso3_all", "eez_sov_iso3_all")) %>%
  mutate(eez_ter_iso3 = ifelse(!is.na(eez_ter_iso3_all), eez_ter_iso3_all, eez_sov_iso3_all)) %>%
  rename(eez_sov_iso3 = eez_sov_iso3_all) %>%
  dplyr::select(-eez_ter_iso3_all) %>%
  dplyr::filter(!is.na(eez_ter_iso3)) %>%
  dplyr::filter(!is.na(flag_iso3)) # remove entries where the flag is unknown - NEW 2/1/2022
```

# Flag state / EEZ summaries (Distant Water Fishing)

Since our vessel list of distant water fishing (and subsidies) is currently at the individual vessel level, we first want to aggregate this by flag state/coastal state. We also want to output summary tables here ranking distant water fishing activity by flag state and EEZ. 

```{r}
# Now calculate summaries for each EEZ - these will be used for the connectivity plots. 
# Note: as of 3/30/22 we realized we were overestimating total tonnage and engine power because vessel characteristics for vessels fishing inside and outside of territorial waters and across multiple area types (joint regime, 200nm, overlapping claim) cooresponding to the same administering territory were being double counted. This didn't affect fishing hour and subsidy calculations. 
flag_eez_id_summary_vessel_stats <- vessel_list_dw_expanded %>%
   group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU, ssvid) %>%
   summarize(length_m = unique(length_m),
             tonnage_gt = unique(tonnage_gt),
             engine_power_kw = unique(engine_power_kw)) %>%
  ungroup() %>%
  group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T)) %>%
  ungroup()

flag_eez_id_summary_fishing_stats <- vessel_list_dw_expanded %>%
  group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU) %>%
  summarize(fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T),
            eez_id = list(unique(eez_id))) %>%
  ungroup()

flag_eez_id_summary <- flag_eez_id_summary_vessel_stats %>%
  left_join(flag_eez_id_summary_fishing_stats) %>%
  unnest(eez_id) %>%
  arrange(eez_ter_iso3, eez_id, flag_iso3)

# Create an aggregated entry for the EU EEZ as a whole
eu_eez_id_summary_vessel_stats <- vessel_list_dw_expanded %>%
  dplyr::filter(eez_is_EU) %>%
  mutate(eez_ter_iso3 = "EU") %>%
  group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU, ssvid) %>%
   summarize(length_m = unique(length_m),
             tonnage_gt = unique(tonnage_gt),
             engine_power_kw = unique(engine_power_kw)) %>%
  ungroup() %>%
  group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T)) %>%
  ungroup()
  
eu_eez_id_summary_fishing_stats <- vessel_list_dw_expanded %>%
  dplyr::filter(eez_is_EU) %>%
  mutate(eez_ter_iso3 = "EU") %>%  
  group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU) %>%
  summarize(fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup() 

eu_eez_id_summary <- eu_eez_id_summary_vessel_stats %>%
  left_join(eu_eez_id_summary_fishing_stats) %>%
  arrange(eez_ter_iso3, flag_iso3) %>%
  mutate(eez_id = 9999)

# Join back together
flag_eez_id_summary <- flag_eez_id_summary %>%
  bind_rows(eu_eez_id_summary)

write_csv(flag_eez_id_summary, paste0(results_dir, "flag_eez_id_summary.csv"))
write_csv(flag_eez_id_summary, paste0(app_dir, "flag_eez_id_summary.csv"))
```

```{r}
# Now we do the same by coastal state... slightly different - these will be used for the summary tables.  
flag_eez_ter_summary <- vessel_list_dw_expanded %>%
  group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T),
            eez_id_all = paste0(unique(eez_id), collapse = ",")) %>%
  ungroup() %>%
  arrange(eez_ter_iso3, flag_iso3)

# Create an aggregated entry for the EU EEZ as a whole
eu_eez_ter_summary <- vessel_list_dw_expanded %>%
  dplyr::filter(eez_is_EU) %>%
  mutate(eez_ter_iso3 = "EU") %>%
  group_by(eez_ter_iso3, eez_is_EU, flag_iso3, flag_is_EU) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup() %>%
  arrange(eez_ter_iso3, flag_iso3) %>%
  mutate(eez_id_all = "9999")

# Join back together
flag_eez_ter_summary <- flag_eez_ter_summary %>%
  bind_rows(eu_eez_ter_summary)

write_csv(flag_eez_ter_summary, paste0(results_dir, "flag_eez_ter_summary.csv"))
write_csv(flag_eez_ter_summary, paste0(app_dir, "flag_eez_ter_summary.csv"))
```

Now let's create a pretty summary version by EEZ ID that will be available for download... 

```{r}
### Let's quickly create some new variables pertaining to all types of areas cooresponding to an administering territory
area_lookup <- vessel_list_dw_expanded %>%
  group_by(eez_ter_iso3) %>%
  mutate(n_eez_areas = n_distinct(eez_id[pol_type == "200NM"]),
         n_overlapping = n_distinct(eez_id[pol_type == "Overlapping claim"]),
         n_joint = n_distinct(eez_id[pol_type == "Joint regime"])) %>%
  ungroup() %>%
  distinct(geoname, eez_id, pol_type, n_eez_areas, n_overlapping, n_joint) %>%
  mutate(mark = case_when(
    pol_type == "200NM" & n_eez_areas > 1 & n_overlapping > 0 & n_joint > 0 ~ "multi-eez & overlapping & joint",
    pol_type == "200NM" & n_eez_areas > 1 & n_overlapping > 0 ~ "multi-eez & overlapping",
    pol_type == "200NM" & n_eez_areas > 1 & n_joint > 0 ~ "multi-eez & joint",
    pol_type == "200NM" & n_overlapping > 0 & n_joint > 0 ~ "overlapping & joint",
    pol_type == "200NM" & n_eez_areas > 1 ~ "multi-eez",
    pol_type == "200NM" & n_overlapping >= 1 ~ "overlapping",
    pol_type == "200NM" & n_joint >= 1 ~ "joint",
    pol_type == "Overlapping claim" ~ "is_overlapping",
    pol_type == "Joint regime" ~ "is_joint",
    TRUE ~ "none"))

eu_area_lookup <- tibble(eez_id = 9999,
                         mark = "aggregate")

area_lookup_out <- area_lookup %>%
  dplyr::select(eez_id, mark) %>%
  bind_rows(eu_area_lookup)

# Now let's create the actual summary characteristics
# Deal with the normal EEZs first
eez_summary <- vessel_list_dw_expanded %>%
  dplyr::filter(pol_type == "200NM") %>%
  group_by(geoname, eez_id) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup()

# We need to do some additional wrangling to the overlapping claim and joint regime areas because we duplicated these earlier
eez_summary_overlapping_joint <- vessel_list_dw_expanded %>%
  dplyr::filter(pol_type %in% c("Overlapping claim", "Joint regime")) %>%
  group_by(eez_id, geoname, ssvid, year, is_territorial, fao_region) %>%
  summarize(length_m = unique(length_m),
            tonnage_gt = unique(tonnage_gt),
            engine_power_kw = unique(engine_power_kw),
            fishing_hours_eez_fao_ter = unique(fishing_hours_eez_fao_ter),
            fishing_KWh_eez_fao_ter = unique(fishing_KWh_eez_fao_ter),
            bad_subs = unique(bad_subs),
            ugly_subs = unique(ugly_subs)) %>%
  ungroup() %>%
  group_by(geoname, eez_id) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup()

# And finally an aggregate entry for the EU
eu_eez_summary <- vessel_list_dw_expanded %>%
  dplyr::filter(eez_is_EU) %>%
  mutate(geoname = "European Union Exclusive Economic Zone",
         eez_id = 9999) %>%
  group_by(geoname, eez_id) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup()

eez_summary <- eez_summary %>%
  bind_rows(eez_summary_overlapping_joint) %>%
  bind_rows(eu_eez_summary) %>%
  arrange(desc(fishing_hours)) %>%
  left_join(area_lookup_out, by = "eez_id")

write_csv(eez_summary, paste0(results_dir, "dw_activity_summary_by_eez.csv"))
write_csv(eez_summary, paste0(app_www_dir, "dw_activity_summary_by_eez.csv"))
```

And then a pretty summary version by flag state for EEZ distant water fishing that will be available for download...

```{r}
flag_state_summary <- vessel_list_dw_expanded %>%
  group_by(flag_iso3) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup() 

flag_state_eu_summary <- vessel_list_dw_expanded %>%
  dplyr::filter(flag_is_EU) %>%
  mutate(flag_iso3 = "EU") %>%
  group_by(flag_iso3) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup()

flag_state_summary <- flag_state_summary %>%
  bind_rows(flag_state_eu_summary) %>%
  arrange(desc(fishing_hours))

write_csv(flag_state_summary, paste0(results_dir, "dw_activity_summary_by_flag_state.csv"))
write_csv(flag_state_summary, paste0(app_www_dir, "dw_activity_summary_by_flag_state.csv"))
```

# Flag state / FAO Region Summaries (Distant Water Fishing)

We now repeat the process for our high seas data, but this is more straightforward because we're aggregating by FAO area and there are no administering territories to deal with.  

```{r}
# Load vessel list
vessel_list_hs_raw <- read.csv(vessel_list_hs_path, stringsAsFactors = F) %>%
  dplyr::filter(!is.na(flag_iso3)) # REMOVE ENTRIES WITH NO KNOWN FLAG STATE - NEW 2/1/2022
```

```{r}
# Get fishing activity by flag for each high seas FAO region
flag_fao_hs_summary <- vessel_list_hs_raw %>%
  group_by(fao_region, flag_iso3, flag_is_EU) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup() %>%
  arrange(fao_region, flag_iso3) 

write_csv(flag_fao_hs_summary, paste0(results_dir, "flag_fao_hs_summary.csv"))
write_csv(flag_fao_hs_summary, paste0(app_dir, "flag_fao_hs_summary.csv"))
```

Now let's create a pretty summary version by high seas FAO area that will be available for download... 

```{r}
fao_summary <- vessel_list_hs_raw %>%
  group_by(fao_region) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup() %>%
  arrange(desc(fishing_hours))

write_csv(fao_summary, paste0(results_dir, "hs_activity_summary_by_area.csv"))
write_csv(fao_summary, paste0(app_www_dir, "hs_activity_summary_by_area.csv"))
```

And finally a pretty summary version by flag state for high seas distant water fishing that will be available for download...

```{r}
hs_flag_state_summary <- vessel_list_hs_raw %>%
  group_by(flag_iso3) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup()

hs_flag_state_eu_summary <- vessel_list_hs_raw %>%
  dplyr::filter(flag_is_EU) %>%
  mutate(flag_iso3 = "EU") %>%
  group_by(flag_iso3) %>%
  summarize(n_vessels = n_distinct(ssvid),
            mean_length = mean(length_m, na.rm = T),
            mean_tonnage = mean(tonnage_gt, na.rm = T),
            tot_tonnage = sum(tonnage_gt, na.rm = T),
            mean_engine_power = mean(engine_power_kw, na.rm = T),
            tot_engine_power = sum(engine_power_kw, na.rm = T),
            fishing_hours = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            ugly_subs = sum(ugly_subs, na.rm = T)) %>%
  ungroup()

hs_flag_state_summary <- hs_flag_state_summary %>%
  bind_rows(hs_flag_state_eu_summary) %>%
  arrange(desc(fishing_hours))

write_csv(hs_flag_state_summary, paste0(results_dir, "hs_activity_summary_by_flag_state.csv"))
write_csv(hs_flag_state_summary, paste0(app_www_dir, "hs_activity_summary_by_flag_state.csv"))
```

# Connectivity Plots (Distant Water Fishing)

## Add centroids to our EEZ id/flag state distant water summary

We first need to load in our land and EEZ centroids so we can start matching them up. 

```{r}
land_centroids <- read_csv(land_centroid_path) 

land_centroids_fix <- land_centroids %>%
  group_by(region, admin_iso3) %>%
  summarize(flag_lon = mean(flag_lon, na.rm = T),
            flag_lat = mean(flag_lat, na.rm = T)) %>%
  ungroup() %>%
  mutate(admin = case_when(admin_iso3 == "EU" ~ "European Union",
                           TRUE ~ countrycode(admin_iso3, "iso3c", "country.name")))

eez_centroids <- read_csv(eez_centroid_path) %>%
  rename(eez_id = mrgid)

eez_eu_centroid <- eez_centroids %>%
  dplyr::filter(eez_ter_iso3 %in% eu_countries & pol_type == "200NM") %>%
  mutate(eez_ter_iso3 = "EU", eez_sov_iso3 = "EU") %>%
  group_by(eez_ter_iso3, eez_sov_iso3, pol_type) %>%
  summarize(eez_lon = mean(eez_lon, na.rm = T),
            eez_lat = mean(eez_lat, na.rm = T)) %>%
  ungroup() %>%
  mutate(eez_id = 9999,
         region = "Europe & Central Asia",
         geoname = "Exclusive Economic Zone:\nEuropean Union",
         eez_ter_name = "European Union",
         eez_sov_name = "European Union")

eez_centroids <- eez_centroids %>%
  bind_rows(eez_eu_centroid)
```

Then we join them together.

```{r}
flag_eez_summary_centroids <- flag_eez_id_summary %>%
  left_join(eez_centroids, by = c("eez_id", "eez_ter_iso3")) %>%
  dplyr::filter(!is.na(eez_lon)) %>% # remove entries for outdated EEZ codes 
  left_join(land_centroids_fix, by = c("region", "flag_iso3" = "admin_iso3"))

# Manually correct for Tuvalu - for some reason it's not in the land file
flag_eez_summary_centroids$flag_lon[flag_eez_summary_centroids$flag_iso3 == "TUV"] <- 177.6493
flag_eez_summary_centroids$flag_lat[flag_eez_summary_centroids$flag_iso3 == "TUV"] <- -7.1095

```

## Draw Connectivity Lines (East Asia & Pacific)

```{r}
# Let's do a test run with East Asia (-10, 350)
eez_mapping_wlines_east_asia_pacific <- flag_eez_summary_centroids %>%
  ungroup() %>%
  dplyr::filter(!is.na(flag_lon)) %>% # remove NA flagged
  dplyr::filter(region == "East Asia & Pacific") %>%
  dplyr::select(flag_lon, flag_lat, eez_lon, eez_lat) %>% 
  purrr::transpose() %>% 
  purrr::map(~ matrix(flatten_dbl(.), nrow = 2, byrow = TRUE)) %>% 
  purrr::map(st_linestring) %>%
  st_sfc(crs = 4326) %>%
  st_sf(geometry = .) %>% 
  bind_cols(flag_eez_summary_centroids %>% dplyr::filter(!is.na(flag_lon)) %>% dplyr::filter(region == "East Asia & Pacific")) %>%
  dplyr::select(everything(), geometry) %>% 
  st_segmentize(units::set_units(100,km)) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES", "DATELINEOFFSET=10"), quiet = TRUE) %>% 
  st_shift_longitude() %>%
  sf::st_sf(crs = 4326) 

test_plot_east_asia_pacific <- ggplot()+
  # geom_sf(data = eez_sf, fill = NA, color = "grey60", size = 0.1)+
  # geom_sf(data = land_sf, fill = "grey2", color = "grey40", size = 0.1)+
  # geom_sf(data = land_sf %>% dplyr::filter(admin_iso3 %in% test_dat_1$flag_iso3), fill = "darkmagenta", color = "grey40", size = 0.1) + # highlighted flag state (magenta, grey border lines)
  # geom_sf(data = eez_sf %>% dplyr::filter(eez_id %in% test_dat_1$eez_id), fill = "slateblue", alpha = 0.5, color = NA, size = 0.1) + # highlighted EEZs (slate blue)
  geom_sf(data = eez_mapping_wlines_east_asia_pacific, col = "darkgoldenrod", size = 0.25) +
  coord_sf(xlim = c(-10,350), ylim = c(-90,90))+
  scale_x_continuous(expand = c(0,0))+
  scale_y_continuous(expand = c(0,0))

test_plot_east_asia_pacific

```

## Draw Connectivity Lines (All Other Regions)

```{r}
# Europe & Central Asia should be straightforward
eez_mapping_wlines_all_others <- flag_eez_summary_centroids %>%
  ungroup() %>%
  dplyr::filter(!is.na(flag_lon)) %>%
  dplyr::filter(region != "East Asia & Pacific") %>%
  dplyr::select(flag_lon, flag_lat, eez_lon, eez_lat) %>% 
  purrr::transpose() %>% 
  purrr::map(~ matrix(flatten_dbl(.), nrow = 2, byrow = TRUE)) %>% 
  purrr::map(st_linestring) %>%
  st_sfc(crs = 4326) %>%
  st_sf(geometry = .) %>% 
  bind_cols(flag_eez_summary_centroids %>% dplyr::filter(!is.na(flag_lon)) %>% dplyr::filter(region != "East Asia & Pacific")) %>%
  dplyr::select(everything(), geometry) %>% 
  st_segmentize(units::set_units(100,km)) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES", "DATELINEOFFSET=10"), quiet = TRUE) %>% 
  sf::st_sf(crs = 4326) 

test_plot_all_others <- ggplot()+
  # geom_sf(data = eez_sf, fill = NA, color = "grey60", size = 0.1)+
  # geom_sf(data = land_sf, fill = "grey2", color = "grey40", size = 0.1)+
  # geom_sf(data = land_sf %>% dplyr::filter(admin_iso3 %in% test_dat_2$flag_iso3), fill = "darkmagenta", color = "grey40", size = 0.1) + # highlighted flag state (magenta, grey border lines)
  # geom_sf(data = eez_sf %>% dplyr::filter(eez_id %in% test_dat_2$eez_id), fill = "slateblue", alpha = 0.5, color = NA, size = 0.1) + # highlighted EEZs (slate blue)
  geom_sf(data = eez_mapping_wlines_all_others, col = "darkgoldenrod", size = 0.25) +
  coord_sf(xlim = c(-180,180), ylim = c(-90,90))+
  scale_x_continuous(expand = c(0,0))+
  scale_y_continuous(expand = c(0,0))

test_plot_all_others

```

We then join them together and save. 

```{r}
eez_mapping_wlines <- eez_mapping_wlines_east_asia_pacific %>%
  rbind(eez_mapping_wlines_all_others)

st_write(eez_mapping_wlines, paste0(results_dir, "eez_flag_state_connectivity_lines_NEW.gpkg"), delete_layer = T)
st_write(eez_mapping_wlines, paste0(app_dir, "eez_flag_state_connectivity_lines_NEW.gpkg"), delete_layer = T)
```

# Connectivity Plots (High Seas Fishing)

We now need to repeat the process for our high seas regions, defined by the portions of the FAO statistical areas that exist outside EEZs.

## Add centroids to our FAO region/flag state distant water summary

We first need to load in our FAO region centroids so we can start matching them up. 

```{r}
fao_centroids <- read_csv(fao_centroid_path) 
```

Then we match them up

```{r}
flag_fao_hs_summary_centroids <- flag_fao_hs_summary %>%
  left_join(fao_centroids, by = c("fao_region" = "zone")) %>%
  dplyr::filter(!is.na(fao_lon)) %>% # removes 1 entry for a russian vessel fishing in the MED
  left_join(land_centroids_fix %>% dplyr::filter(region == "Europe & Central Asia") %>% dplyr::select(-region), by = c("flag_iso3" = "admin_iso3"))

# Manually correct for Tuvalu - for some reason it's not in the land file
flag_fao_hs_summary_centroids$flag_lon[flag_fao_hs_summary_centroids$flag_iso3 == "TUV"] <- 177.6493
flag_fao_hs_summary_centroids$flag_lat[flag_fao_hs_summary_centroids$flag_iso3 == "TUV"] <- -7.1095

```

## Draw Connectivity Lines (High Seas)

```{r}
# Europe & Central Asia should be straightforward
fao_mapping_wlines <- flag_fao_hs_summary_centroids %>%
  ungroup() %>%
  dplyr::filter(!is.na(flag_lon)) %>%
  dplyr::filter(region == "High Seas") %>%
  dplyr::select(flag_lon, flag_lat, fao_lon, fao_lat) %>% 
  purrr::transpose() %>% 
  purrr::map(~ matrix(flatten_dbl(.), nrow = 2, byrow = TRUE)) %>% 
  purrr::map(st_linestring) %>%
  st_sfc(crs = 4326) %>%
  st_sf(geometry = .) %>% 
  bind_cols(flag_fao_hs_summary_centroids %>% dplyr::filter(!is.na(flag_lon)) %>% dplyr::filter(region == "High Seas")) %>%
  dplyr::select(everything(), geometry) %>% 
  st_segmentize(units::set_units(100,km)) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES", "DATELINEOFFSET=10"), quiet = TRUE) %>% 
  sf::st_sf(crs = 4326) 

test_plot_fao <- ggplot()+
  # geom_sf(data = eez_sf, fill = NA, color = "grey60", size = 0.1)+
  # geom_sf(data = land_sf, fill = "grey2", color = "grey40", size = 0.1)+
  # geom_sf(data = land_sf %>% dplyr::filter(admin_iso3 %in% test_dat_2$flag_iso3), fill = "darkmagenta", color = "grey40", size = 0.1) + # highlighted flag state (magenta, grey border lines)
  # geom_sf(data = eez_sf %>% dplyr::filter(eez_id %in% test_dat_2$eez_id), fill = "slateblue", alpha = 0.5, color = NA, size = 0.1) + # highlighted EEZs (slate blue)
  geom_sf(data = fao_mapping_wlines, col = "darkgoldenrod", size = 0.25) +
  coord_sf(xlim = c(-180,180), ylim = c(-90,90))+
  scale_x_continuous(expand = c(0,0))+
  scale_y_continuous(expand = c(0,0))

test_plot_fao

```
 
 We then save. 

```{r}
st_write(fao_mapping_wlines, paste0(results_dir, "fao_flag_state_connectivity_lines.gpkg"), delete_layer = T)
st_write(fao_mapping_wlines, paste0(app_dir, "fao_flag_state_connectivity_lines.gpkg"), delete_layer = T)
```
